move to ModCompatibility
