+++ src/main/kotlin/com/github/noamm9/init/ModCompatibility.kt
@@ -21,4 +21,13 @@ object ModCompatibility {
         blockStateCulling?.setBoolean(config, false)
         mc.levelRenderer.allChanged()
     }
+
+    val bobbyManagesWorlds by lazy {
+        runCatching {
+            if (! isModLoaded("bobby")) return@runCatching false
+            val bobby = Class.forName("de.johni0702.minecraft.bobby.Bobby").getMethod("getInstance").invoke(null)
+            val config = bobby.javaClass.getMethod("getConfig").invoke(bobby)
+            config.javaClass.getMethod("isDynamicMultiWorld").invoke(config) as Boolean
+        }.getOrDefault(false)
+    }
 }
